Split data classes and database into a separate module
