+++ app/src/main/java/ltd/evilcorp/atox/tox/ToxEventListener.kt
@@ -7,15 +7,15 @@ import im.tox.tox4j.core.enums.ToxFileControl
 import im.tox.tox4j.core.enums.ToxMessageType
 import im.tox.tox4j.core.enums.ToxUserStatus
 import ltd.evilcorp.atox.App
-import ltd.evilcorp.atox.repository.ContactRepository
-import ltd.evilcorp.atox.repository.FriendRequestRepository
-import ltd.evilcorp.atox.repository.MessageRepository
-import ltd.evilcorp.atox.repository.UserRepository
 import ltd.evilcorp.atox.ui.NotificationHelper
-import ltd.evilcorp.atox.vo.Contact
-import ltd.evilcorp.atox.vo.FriendRequest
-import ltd.evilcorp.atox.vo.Message
-import ltd.evilcorp.atox.vo.Sender
+import ltd.evilcorp.core.repository.ContactRepository
+import ltd.evilcorp.core.repository.FriendRequestRepository
+import ltd.evilcorp.core.repository.MessageRepository
+import ltd.evilcorp.core.repository.UserRepository
+import ltd.evilcorp.core.vo.Contact
+import ltd.evilcorp.core.vo.FriendRequest
+import ltd.evilcorp.core.vo.Message
+import ltd.evilcorp.core.vo.Sender
 import java.text.DateFormat
 import java.util.*
 import javax.inject.Inject
