+++ app/src/main/java/ltd/evilcorp/atox/ui/contactlist/ContactListViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import ltd.evilcorp.atox.App
-import ltd.evilcorp.atox.repository.ContactRepository
-import ltd.evilcorp.atox.repository.FriendRequestRepository
-import ltd.evilcorp.atox.repository.UserRepository
-import ltd.evilcorp.atox.vo.Contact
-import ltd.evilcorp.atox.vo.FriendRequest
-import ltd.evilcorp.atox.vo.User
+import ltd.evilcorp.core.repository.ContactRepository
+import ltd.evilcorp.core.repository.FriendRequestRepository
+import ltd.evilcorp.core.repository.UserRepository
+import ltd.evilcorp.core.vo.Contact
+import ltd.evilcorp.core.vo.FriendRequest
+import ltd.evilcorp.core.vo.User
 import javax.inject.Inject
 
 class ContactListViewModel @Inject constructor(
