+++ build.gradle
@@ -17,6 +17,16 @@ allprojects {
     }
 }
 
+buildscript {
+    ext {
+        coroutines_version = '1.2.1'
+        dagger_version = '2.23.1'
+        lifecycle_version = '2.2.0-alpha02'
+        room_version = '2.1.0'
+        navigation_version = '2.1.0-beta02'
+    }
+}
+
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
