+++ core/src/main/java/ltd/evilcorp/core/db/UserDao.kt
@@ -1,9 +1,9 @@
-package ltd.evilcorp.atox.db
+package ltd.evilcorp.core.db
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
-import ltd.evilcorp.atox.vo.ConnectionStatus
-import ltd.evilcorp.atox.vo.User
+import ltd.evilcorp.core.vo.ConnectionStatus
+import ltd.evilcorp.core.vo.User
 
 @Dao
 interface UserDao {
